requirements.txt
django
djangorestframework
pillow
psycopg2-binary
drf-yasg
django-filter
dj-rest-auth

